pyproject.toml
[tool.poetry]
name = "sportorg"
version = "1.7.1"
description = "SportOrg, python, sportident, orienteering"
authors = ["Danil Akhtarov <daxartio@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8.1"
pywin32 = {version="^228", optional = true}
pyImpinj = {version="^1.2", optional = true}
playsound = {version="=1.2.2", optional = true}
aiohttp = "^3.9.0"
boltons = "^20"
chardet = "^5.2.0"
docxtpl = "^0"
jinja2 = "^2"
markupsafe = "2.0.1"
orjson = "^3.9.5"
polib = "^1"
psutil = "^5.9.5"
PySide6 = "^6"
python-dateutil = "^2"
pywinusb = "^0"
requests = "^2"
sportident = "^1"
fastapi = "^0.115.12"
pydantic = "<2.0.0"
uvicorn = "<0.18.0"
pyudev = "^0.24.3"
hidapi = "^0.14.0.post4"
pyusb = "1.2.1"

[tool.poetry.group.dev.dependencies]
cx_Freeze = "^6.15.16"
mypy = "^1.4"
poethepoet = ">=0.22,<0.31"
pytest = ">=7.4,<9.0"
pytest-cov = "*"
pytest-deadfixtures = "*"
pytest-mock = "*"
ruff = "^0.8.0"
toml = "*"
types-playsound = "^1.3.1.3"
types-polib = "^1.2.0.1"
types-python-dateutil = "^2.8.19.14"
types-requests = "^2.31.0.2"
types-six = "^1.16.21.9"

[tool.poetry.extras]
win = ["pywin32", "pyImpinj", "playsound"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = [".venv.*"]
ignore_errors = true

[tool.coverage.run]
omit = ["tests/*", "**/__main__.py", "**/.venv/*", "**/site-packages/*"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 42

[tool.pytest.ini_options]
addopts = "--strict-markers --showlocals --verbosity 2"
log_level = "DEBUG"

[tool.ruff.lint]
select = []
ignore = []
exclude = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.poe.env]
CODE = "tests sportorg builder.py"
SPORTORG_DEBUG = "true"

[tool.poe.tasks.all]
help = "Run format lint test"
sequence = ["format", "lint", "test"]
default_item_type = "ref"

[tool.poe.tasks.test]
help = "Test with coverage"
sequence = [
    {ref = "generate-mo"},
    {cmd = "pytest -vv --cov"},
]

[tool.poe.tasks.test-fast]
help = "Test until error"
cmd = "pytest -vv --exitfirst"

[tool.poe.tasks.test-failed]
help = "Test failed"
cmd = "pytest -vv --last-failed"

[tool.poe.tasks.lint]
help = "Check code"
sequence = [
    {cmd = "poetry check"},
    {ref = "generate-mo"},
    {cmd = "ruff check ${CODE}"},
    {cmd = "pytest --dead-fixtures --dup-fixtures"},
    # {cmd = "mypy ${CODE}"},
]

[tool.poe.tasks.format]
help = "Formating code"
sequence = [
    {ref = "generate-mo"},
	{cmd = "ruff format ${CODE}"},
]

[tool.poe.tasks.generate-mo]
help = "Generate mo files"
cmd = "python -m sportorg.language"

[tool.poe.tasks.generate-version]
help = "Generate version file"
cmd = "python .github/scripts/version.py"

[tool.poe.tasks.run]
help = "Run"
sequence = [
    {ref = "generate-mo"},
    {cmd = "python SportOrg.pyw"}
]
